+++ CMakeLists.txt
@@ -125,6 +125,7 @@ option(TILEDB_CCACHE "If true, enables use of 'ccache' (if present)" OFF)
 option(TILEDB_ARROW_TESTS "If true, enables building the arrow adapter unit tests" OFF)
 option(TILEDB_LOG_OUTPUT_ON_FAILURE "If true, print error logs if dependency sub-project build fails" ON)
 option(TILEDB_SKIP_S3AWSSDK_DIR_LENGTH_CHECK "If true, skip check needed path length for awssdk (TILEDB_S3) dependent builds" OFF)
+option(TILEDB_EXPERIMENTAL_FEATURES "If true, build and include experimental features" OFF)
 
 set(TILEDB_INSTALL_LIBDIR "" CACHE STRING "If non-empty, install TileDB library to this directory instead of CMAKE_INSTALL_LIBDIR.")
 
@@ -165,6 +166,11 @@ endif()
 # When building static TileDB, we also need to install any static dependencies
 # built as external projects.
 set(TILEDB_INSTALL_STATIC_DEPS ${TILEDB_STATIC})
+
+# Set main TileDB Source
+set(TILEDB_BASE_SOURCE "${CMAKE_CURRENT_SOURCE_DIR}/tiledb")
+# Set experimental TileDB Source
+set(TILEDB_EXPERIMENTAL_BASE_SOURCE "${CMAKE_CURRENT_SOURCE_DIR}/experimental")
 
 # Perform superbuild config and exit.
 if (TILEDB_SUPERBUILD)
@@ -254,6 +260,9 @@ endif()
 
 # Enable testing
 enable_testing()
+
+# Build the TileDB library experimental features
+add_subdirectory(experimental)
 
 # Build the TileDB library
 add_subdirectory(tiledb)
