+++ bootstrap.ps1
@@ -50,6 +50,9 @@ Enables building TileDB dependencies from source (superbuild)
 .PARAMETER EnableTools
 Enables building TileDB CLI tools (experimental)
 
+.PARAMETER EnableExperimentalFeatures
+Enables building TileDB Experimental features
+
 .Parameter DisableWerror
 Disable use of warnings-as-errors (/WX) during build.
 
@@ -84,6 +87,7 @@ Param(
     [switch]$EnableSerialization,
     [switch]$EnableStaticTileDB,
     [switch]$EnableTools,
+    [switch]$EnableExperimentalFeatures,
     [switch]$EnableBuildDeps,
     [switch]$DisableWerror,
     [switch]$DisableCppApi,
@@ -175,6 +179,11 @@ if ($EnableTools.IsPresent) {
     $TileDBTools = "ON"
 }
 
+$TileDBExperimentalFeatures = "OFF"
+if ($EnableExperimentalFeatures.IsPresent) {
+    $TileDBExperimentalFeatures = "ON"
+}
+
 $TileDBBuildDeps = "OFF";
 if ($EnableBuildDeps.IsPresent) {
     $TileDBBuildDeps = "ON"
@@ -213,7 +222,7 @@ if ($CMakeGenerator -eq $null) {
 
 # Run CMake.
 # We use Invoke-Expression so we can echo the command to the user.
-$CommandString = "cmake -A X64 -DCMAKE_BUILD_TYPE=$BuildType -DCMAKE_INSTALL_PREFIX=""$InstallPrefix"" -DCMAKE_PREFIX_PATH=""$DependencyDir"" -DMSVC_MP_FLAG=""/MP$BuildProcesses"" -DTILEDB_VERBOSE=$Verbosity -DTILEDB_AZURE=$UseAzure -DTILEDB_S3=$UseS3 -DTILEDB_SERIALIZATION=$UseSerialization -DTILEDB_WERROR=$Werror -DTILEDB_CPP_API=$CppApi -DTILEDB_TESTS=$Tests -DTILEDB_STATS=$Stats -DTILEDB_STATIC=$TileDBStatic -DTILEDB_FORCE_ALL_DEPS=$TileDBBuildDeps -DTILEDB_TOOLS=$TileDBTools $GeneratorFlag ""$SourceDirectory"""
+$CommandString = "cmake -A X64 -DCMAKE_BUILD_TYPE=$BuildType -DCMAKE_INSTALL_PREFIX=""$InstallPrefix"" -DCMAKE_PREFIX_PATH=""$DependencyDir"" -DMSVC_MP_FLAG=""/MP$BuildProcesses"" -DTILEDB_VERBOSE=$Verbosity -DTILEDB_AZURE=$UseAzure -DTILEDB_S3=$UseS3 -DTILEDB_SERIALIZATION=$UseSerialization -DTILEDB_WERROR=$Werror -DTILEDB_CPP_API=$CppApi -DTILEDB_TESTS=$Tests -DTILEDB_STATS=$Stats -DTILEDB_STATIC=$TileDBStatic -DTILEDB_FORCE_ALL_DEPS=$TileDBBuildDeps -DTILEDB_TOOLS=$TileDBTools -DTILEDB_EXPERIMENTAL_FEATURES=$TileDBExperimentalFeatures $GeneratorFlag ""$SourceDirectory"""
 Write-Host $CommandString
 Write-Host
 Invoke-Expression "$CommandString"
