+++ cmake/TileDB-Superbuild.cmake
@@ -44,6 +44,7 @@ set(INHERITED_CMAKE_ARGS
   -DTILEDB_ARROW_TESTS=${TILEDB_ARROW_TESTS}
   -DTILEDB_INSTALL_LIBDIR=${TILEDB_INSTALL_LIBDIR}
   -DCMAKE_OSX_ARCHITECTURES=${CMAKE_OSX_ARCHITECTURES}
+  -DTILEDB_EXPERIMENTAL_FEATURES=${TILEDB_EXPERIMENTAL_FEATURES}
 )
 
 if (TILEDB_TESTS)
