+++ experimental/tiledb/sm/array_schema/CMakeLists.txt
@@ -1,0 +1,40 @@
+#
+# experimental/tiledb/sm/array_schema/CMakeLists.txt
+#
+#
+# The MIT License
+#
+# Copyright (c) 2021 TileDB, Inc.
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+#
+
+include(common NO_POLICY_SCOPE)
+
+list(APPEND SOURCES
+        array_schema_evolution.cc
+)
+gather_sources(${SOURCES})
+
+list(APPEND DEPENDENT_SOURCES
+        "${TILEDB_BASE_SOURCE}/sm/array_schema/array_schema.cc"
+        "${TILEDB_BASE_SOURCE}/sm/array_schema/attribute.cc"
+        "${TILEDB_BASE_SOURCE}/sm/array_schema/dimension.cc"
+        "${TILEDB_BASE_SOURCE}/sm/array_schema/domain.cc"
+)
