+++ external/blosc/CMakeLists.txt
@@ -61,12 +61,12 @@ endif()
 # in case something else has a pthread dependency.
 #
 find_package(Threads REQUIRED)
-message("Threads_FOUND=${Threads_FOUND}")
-message("CMAKE_THREAD_LIBS_INIT=${CMAKE_THREAD_LIBS_INIT}")
-message("CMAKE_USE_WIN32_THREADS_INIT=${CMAKE_USE_WIN32_THREADS_INIT}")
-message("CMAKE_USE_PTHREADS_INIT=${CMAKE_USE_PTHREADS_INIT}")
-message("CMAKE_HP_PTHREADS_INIT=${CMAKE_HP_PTHREADS_INIT}")
-message("CMAKE_HAVE_PTHREAD_H=${CMAKE_HAVE_PTHREAD_H}")
+#message("Threads_FOUND=${Threads_FOUND}")
+#message("CMAKE_THREAD_LIBS_INIT=${CMAKE_THREAD_LIBS_INIT}")
+#message("CMAKE_USE_WIN32_THREADS_INIT=${CMAKE_USE_WIN32_THREADS_INIT}")
+#message("CMAKE_USE_PTHREADS_INIT=${CMAKE_USE_PTHREADS_INIT}")
+#message("CMAKE_HP_PTHREADS_INIT=${CMAKE_HP_PTHREADS_INIT}")
+#message("CMAKE_HAVE_PTHREAD_H=${CMAKE_HAVE_PTHREAD_H}")
 
 if (CMAKE_USE_PTHREADS_INIT)
   message("Using system-provided pthread")
