+++ tiledb/sm/array/array.cc
@@ -31,9 +31,9 @@
  */
 
 #include "tiledb/sm/array/array.h"
+#include "experimental/tiledb/sm/array_schema/array_schema_evolution.h"
 #include "tiledb/common/logger.h"
 #include "tiledb/sm/array_schema/array_schema.h"
-#include "tiledb/sm/array_schema/array_schema_evolution.h"
 #include "tiledb/sm/array_schema/attribute.h"
 #include "tiledb/sm/array_schema/dimension.h"
 #include "tiledb/sm/array_schema/domain.h"
