+++ tiledb/sm/c_api/tiledb.cc
@@ -32,11 +32,11 @@
  */
 
 #include "tiledb/sm/c_api/tiledb.h"
+#include "experimental/tiledb/sm/serialization/array_schema_evolution.h"
 #include "tiledb/common/heap_profiler.h"
 #include "tiledb/common/logger.h"
 #include "tiledb/sm/array/array.h"
 #include "tiledb/sm/array_schema/array_schema.h"
-#include "tiledb/sm/c_api/tiledb_experimental.h"
 #include "tiledb/sm/c_api/api_validation.h"
 #include "tiledb/sm/c_api/tiledb_serialization.h"
 #include "tiledb/sm/c_api/tiledb_struct_def.h"
@@ -62,7 +62,6 @@
 #include "tiledb/sm/query/query_condition.h"
 #include "tiledb/sm/rest/rest_client.h"
 #include "tiledb/sm/serialization/array_schema.h"
-#include "tiledb/sm/serialization/array_schema_evolution.h"
 #include "tiledb/sm/serialization/config.h"
 #include "tiledb/sm/serialization/query.h"
 #include "tiledb/sm/stats/global_stats.h"
@@ -2449,91 +2448,6 @@ int32_t tiledb_array_schema_has_attribute(
   return TILEDB_OK;
 }
 
-/* ********************************* */
-/*            SCHEMA EVOLUTION       */
-/* ********************************* */
-
-int32_t tiledb_array_schema_evolution_alloc(
-    tiledb_ctx_t* ctx,
-    tiledb_array_schema_evolution_t** array_schema_evolution) {
-  // Sanity check
-  if (sanity_check(ctx) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  // Create schema evolution struct
-  *array_schema_evolution = new (std::nothrow) tiledb_array_schema_evolution_t;
-  if (*array_schema_evolution == nullptr) {
-    auto st = Status::Error(
-        "Failed to allocate TileDB array schema evolution object");
-    LOG_STATUS(st);
-    save_error(ctx, st);
-    return TILEDB_OOM;
-  }
-
-  // Create a new SchemaEvolution object
-  (*array_schema_evolution)->array_schema_evolution_ =
-      new (std::nothrow) tiledb::sm::ArraySchemaEvolution();
-  if ((*array_schema_evolution)->array_schema_evolution_ == nullptr) {
-    delete *array_schema_evolution;
-    *array_schema_evolution = nullptr;
-    auto st = Status::Error(
-        "Failed to allocate TileDB array schema evolution object");
-    LOG_STATUS(st);
-    save_error(ctx, st);
-    return TILEDB_OOM;
-  }
-
-  // Success
-  return TILEDB_OK;
-}
-
-void tiledb_array_schema_evolution_free(
-    tiledb_array_schema_evolution_t** array_schema_evolution) {
-  if (array_schema_evolution != nullptr && *array_schema_evolution != nullptr) {
-    delete (*array_schema_evolution)->array_schema_evolution_;
-    delete *array_schema_evolution;
-    *array_schema_evolution = nullptr;
-  }
-}
-
-int32_t tiledb_array_schema_evolution_add_attribute(
-    tiledb_ctx_t* ctx,
-    tiledb_array_schema_evolution_t* array_schema_evolution,
-    tiledb_attribute_t* attr) {
-  if (sanity_check(ctx) == TILEDB_ERR ||
-      sanity_check(ctx, array_schema_evolution) == TILEDB_ERR ||
-      sanity_check(ctx, attr) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  if (SAVE_ERROR_CATCH(
-          ctx,
-          array_schema_evolution->array_schema_evolution_->add_attribute(
-              attr->attr_)))
-    return TILEDB_ERR;
-  return TILEDB_OK;
-
-  // Success
-  return TILEDB_OK;
-}
-
-int32_t tiledb_array_schema_evolution_drop_attribute(
-    tiledb_ctx_t* ctx,
-    tiledb_array_schema_evolution_t* array_schema_evolution,
-    const char* attribute_name) {
-  if (sanity_check(ctx) == TILEDB_ERR ||
-      sanity_check(ctx, array_schema_evolution) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  if (SAVE_ERROR_CATCH(
-          ctx,
-          array_schema_evolution->array_schema_evolution_->drop_attribute(
-              attribute_name)))
-    return TILEDB_ERR;
-  return TILEDB_OK;
-  // Success
-  return TILEDB_OK;
-}
-
 /* ****************************** */
 /*              QUERY             */
 /* ****************************** */
@@ -4457,71 +4371,6 @@ int32_t tiledb_array_consolidate_metadata_with_key(
   return TILEDB_OK;
 }
 
-int32_t tiledb_array_evolve(
-    tiledb_ctx_t* ctx,
-    const char* array_uri,
-    tiledb_array_schema_evolution_t* array_schema_evolution) {
-  // Sanity Checks
-  if (sanity_check(ctx) == TILEDB_ERR ||
-      sanity_check(ctx, array_schema_evolution) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  // Check array name
-  tiledb::sm::URI uri(array_uri);
-  if (uri.is_invalid()) {
-    auto st = Status::Error("Failed to create array; Invalid array URI");
-    LOG_STATUS(st);
-    save_error(ctx, st);
-    return TILEDB_ERR;
-  }
-
-  // Create key
-  tiledb::sm::EncryptionKey key;
-  if (SAVE_ERROR_CATCH(
-          ctx,
-          key.set_key(
-              static_cast<tiledb::sm::EncryptionType>(TILEDB_NO_ENCRYPTION),
-              nullptr,
-              0)))
-    return TILEDB_ERR;
-
-  if (SAVE_ERROR_CATCH(
-          ctx,
-          ctx->ctx_->storage_manager()->array_evolve_schema(
-              uri, array_schema_evolution->array_schema_evolution_, key)))
-    return TILEDB_ERR;
-
-  // Success
-  return TILEDB_OK;
-}
-
-int32_t tiledb_array_upgrade_version(
-    tiledb_ctx_t* ctx, const char* array_uri, tiledb_config_t* config) {
-  // Sanity Checks
-  if (sanity_check(ctx) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  // Check array name
-  tiledb::sm::URI uri(array_uri);
-  if (uri.is_invalid()) {
-    auto st = Status::Error("Failed to find the array; Invalid array URI");
-    LOG_STATUS(st);
-    save_error(ctx, st);
-    return TILEDB_ERR;
-  }
-
-  if (SAVE_ERROR_CATCH(
-          ctx,
-          ctx->ctx_->storage_manager()->array_upgrade_version(
-              uri,
-              (config == nullptr) ? &ctx->ctx_->storage_manager()->config() :
-                                    config->config_)))
-    return TILEDB_ERR;
-
-  // Success
-  return TILEDB_OK;
-}
-
 /* ****************************** */
 /*         OBJECT MANAGEMENT      */
 /* ****************************** */
@@ -5310,74 +5159,6 @@ int32_t tiledb_deserialize_array_schema(
               *buffer->buffer_))) {
     delete *array_schema;
     *array_schema = nullptr;
-    return TILEDB_ERR;
-  }
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_serialize_array_schema_evolution(
-    tiledb_ctx_t* ctx,
-    const tiledb_array_schema_evolution_t* array_schema_evolution,
-    tiledb_serialization_type_t serialize_type,
-    int32_t client_side,
-    tiledb_buffer_t** buffer) {
-  // Sanity check
-  if (sanity_check(ctx) == TILEDB_ERR ||
-      sanity_check(ctx, array_schema_evolution) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  // Create buffer
-  if (tiledb_buffer_alloc(ctx, buffer) != TILEDB_OK ||
-      sanity_check(ctx, *buffer) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  if (SAVE_ERROR_CATCH(
-          ctx,
-          tiledb::sm::serialization::array_schema_evolution_serialize(
-              array_schema_evolution->array_schema_evolution_,
-              (tiledb::sm::SerializationType)serialize_type,
-              (*buffer)->buffer_,
-              client_side))) {
-    tiledb_buffer_free(buffer);
-    return TILEDB_ERR;
-  }
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_deserialize_array_schema_evolution(
-    tiledb_ctx_t* ctx,
-    const tiledb_buffer_t* buffer,
-    tiledb_serialization_type_t serialize_type,
-    int32_t client_side,
-    tiledb_array_schema_evolution_t** array_schema_evolution) {
-  // Currently unused:
-  (void)client_side;
-
-  // Sanity check
-  if (sanity_check(ctx) == TILEDB_ERR ||
-      sanity_check(ctx, buffer) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  // Create array schema struct
-  *array_schema_evolution = new (std::nothrow) tiledb_array_schema_evolution_t;
-  if (*array_schema_evolution == nullptr) {
-    auto st = Status::Error(
-        "Failed to allocate TileDB array schema evolution object");
-    LOG_STATUS(st);
-    save_error(ctx, st);
-    return TILEDB_OOM;
-  }
-
-  if (SAVE_ERROR_CATCH(
-          ctx,
-          tiledb::sm::serialization::array_schema_evolution_deserialize(
-              &((*array_schema_evolution)->array_schema_evolution_),
-              (tiledb::sm::SerializationType)serialize_type,
-              *buffer->buffer_))) {
-    delete *array_schema_evolution;
-    *array_schema_evolution = nullptr;
     return TILEDB_ERR;
   }
